reorganize default os pool
