+++ src/orch/thread_pool/implementations/mod.rs
@@ -1,3 +1,4 @@
-mod std_default_thread_pool;
+mod std_os_thread_pool;
+mod std_scoped_threads;
 
-pub use std_default_thread_pool::StdDefaultThreadPool;
+pub use std_os_thread_pool::StdOsThreadPool;
